+++ src/widgets/settingspages/ModerationPage.cpp
@@ -81,16 +81,16 @@ ModerationPage::ModerationPage()
             "Enable logging", getSettings()->enableLogging);
         logs.append(enableLogging);
 
-		auto description = logs.emplace<QLabel>(
-				"Logging saves the chat history in plain text "
-				"with time and date and emote codes.");
-		description->setStyleSheet("color: #bbb");
-
         auto logsPathLabel = logs.emplace<QLabel>();
+
+		QString logExplanation =
+			"<span style=\"color:#bbb\"> They are saved as plain "
+			"text files per channel, containing the messages with "
+			"timestamps.</span>";
 
         // Logs (copied from LoggingMananger)
         getSettings()->logPath.connect(
-            [logsPathLabel](const QString &logPath, auto) mutable {
+            [logsPathLabel, logDescription](const QString &logPath, auto) mutable {
                 QString pathOriginal =
                     logPath.isEmpty() ? getApp()->getPaths().messageLogDirectory
                                       : logPath;
@@ -98,10 +98,11 @@ ModerationPage::ModerationPage()
                 QString pathShortened =
                     "Logs are saved at <a href=\"file:///" + pathOriginal +
                     R"("><span style="color: white;">)" +
-                    shortenString(pathOriginal, 50) + "</span></a>";
+                    shortenString(pathOriginal, 50) + ".</span></a>";
 
-                logsPathLabel->setText(pathShortened);
+                logsPathLabel->setText(pathShortened + logExplanation);
                 logsPathLabel->setToolTip(pathOriginal);
+				logsPathLabel->setWordWrap(true);
             });
 
         logsPathLabel->setTextFormat(Qt::RichText);